+++ tests/testthat/test-capping.R
@@ -15,6 +15,11 @@ test_that("irace_capping_xy maxExperiments = 1000", {
   irace_capping_xy(maxExperiments = 1000)
 })
 
+test_that("irace_capping_xy maxExperiments = 1000 cappingAfterFirstTest", {
+  generate_set_seed()
+  irace_capping_xy(maxExperiments = 1000, cappingAfterFirstTest=1)
+})
+
 test_that("irace_capping_xy maxTime = 1000", {
   generate_set_seed()
   expect_warning(irace_capping_xy(maxTime = 1000),
@@ -22,7 +27,7 @@ test_that("irace_capping_xy maxTime = 1000", {
 })
 
 test_that("irace_capping_xy targetRunner = target_runner_reject, maxTime = 1000", {
-  skip_on_cran() 
+  skip_on_cran()
   skip_on_coverage() # This test sometimes fails randomly
   generate_set_seed()
   irace_capping_xy(targetRunner = target_runner_reject, maxTime = 1000, boundMax = 5, debugLevel = 3,
@@ -35,7 +40,7 @@ test_that("capping default value", {
    x "" r (0, 1.00)
    y "" r (0, 1.00)
    reject "" c (0,1)'
-  
+
   parameters <- readParameters(text = parameters_table)
 
   def_scenario <- list(instances = c("ackley", "goldestein", "matyas", "himmelblau"),